renamed input field file and added xml docs
