+++ Satchel.csproj
@@ -109,7 +109,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BetterMenus\Elements\NumInputField.cs" />
+    <Compile Include="BetterMenus\Elements\InputField.cs" />
     <Compile Include="GlobalUsings.cs" />
     <Compile Include="Mod.cs" />
     <Compile Include="Core.cs" />
